made some simplistic vars to handle screen width/height  in params for objects / methods. fixed scaling for bg's part 3.

+++ app/src/main/java/com/example/hyperponggruppb/view/StoryView.kt
@@ -96,8 +96,8 @@ class StoryView(var myContext: Context?, var activity: Activity) : SurfaceView(m
 
             canvas = mHolder!!.lockCanvas()
 
-            PsyduckEngine.canvasHeight = canvas.height.toFloat()
-            PsyduckEngine.canvasWidth = canvas.width.toFloat()
+            PsyduckEngine.canvasHeight = canvas.height
+            PsyduckEngine.canvasWidth = canvas.width
 
             canvas.drawBitmap(AssetManager.getBackground(backgroundCode), matrix, null)
